Fixed issue getting files from root S3 directory
Resolved an issue where if you tried to fetch files from the root
directory of an S3 bucket you would get nothing back. We were always
setting the prefix to '/' which is invalid, and therefore returned
nothing.

Added test cases for both readdir() and sync() both on S3 and local
storage. There is another bug which resulted from this, where if you
name consecutive directories the same thing the file will not be synced
down.